Refactor and locally test backend
