+++ backend/backend/.gitignore
@@ -43,3 +43,5 @@ local_settings.py
 *.sublime-workspace
 *.sublime-project
 *.sql
+
+data/
