+++ backend/backend/api/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from . import views
+from .views import views
 
 urlpatterns = [
     path('get/', views.my_view),
