+++ backend/backend/requirements.txt
@@ -1,3 +1,5 @@
 asgiref==3.8.1
 Django==5.1.2
 sqlparse==0.5.1
+pandas
+elasticsearch
