+++ backend/readme.md
@@ -1,1 +1,4 @@
-python manage.py runserver
+### Python 3.13
+
+python3 -m pip install -r requirements.txt
+python3 manage.py runserver
